feat: Add prompt

+++ scripts/uninstall.sh
@@ -80,18 +80,20 @@ main() {
 
   # Remove Spaceship from .zshrc
   if grep -q "spaceship" "$ZSHRC"; then
-    info "Removing Spaceship from ~/.zshrc"
-    # Remove enabling statements from ~/.zshrc
-    # and remove Spaceship configuration
-    # Note: SPACESHIP_RPROMPT_ORDER and SPACESHIP_PROMPT_ORDER configuration may have multiple lines
-    # which are grouped by `(`, `)`
-    sed '/^# Set Spaceship ZSH as a prompt$/d' "$ZSHRC" | \
-    sed '/^autoload -U promptinit; promptinit$/d' | \
-    sed '/^prompt spaceship$/d' | \
-    sed  -E '/^SPACESHIP_R?PROMPT_ORDER=\([^)]*$/,/^[^(]*)/d' | \
-    sed '/^SPACESHIP_.*$/d' > "$ZSHRC.bak"
-
-    mv -- "$ZSHRC.bak" "$ZSHRC"
+    read yesOrNo'?Would you like to remove you Spaceship ZSH configuration from .zshrc? (y/N)'
+    if [[ 'y' == ${yesOrNo:l} ]]; then
+      info "Removing Spaceship from ~/.zshrc"
+      # Remove enabling statements from ~/.zshrc
+      # and remove Spaceship configuration
+      # Note: SPACESHIP_RPROMPT_ORDER and SPACESHIP_PROMPT_ORDER configuration may have multiple lines
+      # which are grouped by `(`, `)`
+      sed '/^# Set Spaceship ZSH as a prompt$/d' "$ZSHRC" | \
+      sed '/^autoload -U promptinit; promptinit$/d' | \
+      sed '/^prompt spaceship$/d' | \
+      sed  -E '/^SPACESHIP_R?PROMPT_ORDER=\([^)]*$/,/^[^(]*)/d' | \
+      sed '/^SPACESHIP_.*$/d' > "$ZSHRC.bak" && \
+      mv -- "$ZSHRC.bak" "$ZSHRC"
+    fi
   else
     warn "Spaceship configuration not found in ~/.zshrc!"
   fi
